+++ src/components/Tasks/TasksContext.tsx
@@ -4,8 +4,8 @@ import { useTasks } from '../../hooks/useTasks'
 export type TasksContextType = ReturnType<typeof useTasks>
 export const TasksContext = createContext<TasksContextType>({
   data: undefined,
-  isLoading: undefined,
-  isValidating: undefined,
+  isLoading: false,
+  isValidating: false,
   error: undefined,
   refetch: () => {},
   submitResponseToTask: () => {},
