+++ src/styles/typography.scss
@@ -82,7 +82,7 @@
 .Layer__text[data-status='error'] {
   color: var(--color-danger);
 
-  &[data-invert='true'] {
+  &[data-inverted] {
     color: var(--color-info-error);
   }
 }
@@ -90,7 +90,7 @@
 .Layer__text[data-status='warning'] {
   color: var(--color-info-warning-fg);
 
-  &[data-invert='true'] {
+  &[data-inverted] {
     color: var(--color-info-warning-bg);
   }
 }
@@ -98,7 +98,7 @@
 .Layer__text[data-status='success'] {
   color: var(--color-success);
 
-  &[data-invert='true'] {
+  &[data-inverted] {
     color: var(--color-info-success-bg);
   }
 }
@@ -110,7 +110,7 @@
 .Layer__text[data-status='info'] {
   color: var(--color-info-fg);
 
-  &[data-invert='true'] {
+  &[data-inverted] {
     color: var(--color-info-bg);
   }
 }
